+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/FunctionsService.java
@@ -20,6 +20,8 @@ public interface FunctionsService {
   /**
    * Create a function.
    *
+   * <p>**WARNING: This API is experimental and will change in future versions**
+   *
    * <p>Creates a new function
    *
    * <p>The user must have the following permissions in order for the function to be created: -