+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/ListCatalogsRequest.java
@@ -17,6 +17,22 @@ public class ListCatalogsRequest {
   @QueryParam("include_browse")
   private Boolean includeBrowse;
 
+  /**
+   * Maximum number of catalogs to return. - when set to 0, the page length is set to a server
+   * configured value (recommended); - when set to a value greater than 0, the page length is the
+   * minimum of this value and a server configured value; - when set to a value less than 0, an
+   * invalid parameter error is returned; - If not set, all valid catalogs are returned (not
+   * recommended). - Note: The number of returned catalogs might be less than the specified
+   * max_results size, even zero. The only definitive indication that no further catalogs can be
+   * fetched is when the next_page_token is unset from the response.
+   */
+  @QueryParam("max_results")
+  private Long maxResults;
+
+  /** Opaque pagination token to go to next page based on previous query. */
+  @QueryParam("page_token")
+  private String pageToken;
+
   public ListCatalogsRequest setIncludeBrowse(Boolean includeBrowse) {
     this.includeBrowse = includeBrowse;
     return this;
@@ -26,21 +42,45 @@ public class ListCatalogsRequest {
     return includeBrowse;
   }
 
+  public ListCatalogsRequest setMaxResults(Long maxResults) {
+    this.maxResults = maxResults;
+    return this;
+  }
+
+  public Long getMaxResults() {
+    return maxResults;
+  }
+
+  public ListCatalogsRequest setPageToken(String pageToken) {
+    this.pageToken = pageToken;
+    return this;
+  }
+
+  public String getPageToken() {
+    return pageToken;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
     ListCatalogsRequest that = (ListCatalogsRequest) o;
-    return Objects.equals(includeBrowse, that.includeBrowse);
+    return Objects.equals(includeBrowse, that.includeBrowse)
+        && Objects.equals(maxResults, that.maxResults)
+        && Objects.equals(pageToken, that.pageToken);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(includeBrowse);
+    return Objects.hash(includeBrowse, maxResults, pageToken);
   }
 
   @Override
   public String toString() {
-    return new ToStringer(ListCatalogsRequest.class).add("includeBrowse", includeBrowse).toString();
+    return new ToStringer(ListCatalogsRequest.class)
+        .add("includeBrowse", includeBrowse)
+        .add("maxResults", maxResults)
+        .add("pageToken", pageToken)
+        .toString();
   }
 }