+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/UpdateStorageCredential.java
@@ -29,13 +29,20 @@ public class UpdateStorageCredential {
   @JsonProperty("comment")
   private String comment;
 
-  /** The <Databricks> managed GCP service account configuration. */
+  /** The Databricks managed GCP service account configuration. */
   @JsonProperty("databricks_gcp_service_account")
   private DatabricksGcpServiceAccountRequest databricksGcpServiceAccount;
 
   /** Force update even if there are dependent external locations or external tables. */
   @JsonProperty("force")
   private Boolean force;
+
+  /**
+   * Whether the current securable is accessible from all workspaces or a specific set of
+   * workspaces.
+   */
+  @JsonProperty("isolation_mode")
+  private IsolationMode isolationMode;
 
   /** Name of the storage credential. */
   private String name;
@@ -122,6 +129,15 @@ public class UpdateStorageCredential {
     return force;
   }
 
+  public UpdateStorageCredential setIsolationMode(IsolationMode isolationMode) {
+    this.isolationMode = isolationMode;
+    return this;
+  }
+
+  public IsolationMode getIsolationMode() {
+    return isolationMode;
+  }
+
   public UpdateStorageCredential setName(String name) {
     this.name = name;
     return this;
@@ -179,6 +195,7 @@ public class UpdateStorageCredential {
         && Objects.equals(comment, that.comment)
         && Objects.equals(databricksGcpServiceAccount, that.databricksGcpServiceAccount)
         && Objects.equals(force, that.force)
+        && Objects.equals(isolationMode, that.isolationMode)
         && Objects.equals(name, that.name)
         && Objects.equals(newName, that.newName)
         && Objects.equals(owner, that.owner)
@@ -196,6 +213,7 @@ public class UpdateStorageCredential {
         comment,
         databricksGcpServiceAccount,
         force,
+        isolationMode,
         name,
         newName,
         owner,
@@ -213,6 +231,7 @@ public class UpdateStorageCredential {
         .add("comment", comment)
         .add("databricksGcpServiceAccount", databricksGcpServiceAccount)
         .add("force", force)
+        .add("isolationMode", isolationMode)
         .add("name", name)
         .add("newName", newName)
         .add("owner", owner)