+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/marketplace/Listing.java
@@ -17,6 +17,13 @@ public class Listing {
   @JsonProperty("id")
   private String id;
 
+  /**
+   * we can not use just ProviderListingSummary since we already have same name on entity side of
+   * the state
+   */
+  @JsonProperty("provider_summary")
+  private ProviderListingSummaryInfo providerSummary;
+
   /** Next Number: 26 */
   @JsonProperty("summary")
   private ListingSummary summary;
@@ -39,6 +46,15 @@ public class Listing {
     return id;
   }
 
+  public Listing setProviderSummary(ProviderListingSummaryInfo providerSummary) {
+    this.providerSummary = providerSummary;
+    return this;
+  }
+
+  public ProviderListingSummaryInfo getProviderSummary() {
+    return providerSummary;
+  }
+
   public Listing setSummary(ListingSummary summary) {
     this.summary = summary;
     return this;
@@ -55,12 +71,13 @@ public class Listing {
     Listing that = (Listing) o;
     return Objects.equals(detail, that.detail)
         && Objects.equals(id, that.id)
+        && Objects.equals(providerSummary, that.providerSummary)
         && Objects.equals(summary, that.summary);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(detail, id, summary);
+    return Objects.hash(detail, id, providerSummary, summary);
   }
 
   @Override
@@ -68,6 +85,7 @@ public class Listing {
     return new ToStringer(Listing.class)
         .add("detail", detail)
         .add("id", id)
+        .add("providerSummary", providerSummary)
         .add("summary", summary)
         .toString();
   }