+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sql/AlertsService.java
@@ -10,6 +10,11 @@ import java.util.Collection;
  * more users and/or notification destinations if the condition was met. Alerts can be scheduled
  * using the `sql_task` type of the Jobs API, e.g. :method:jobs/create.
  *
+ * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+ *
+ * <p>[Learn more]:
+ * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
+ *
  * <p>This is the high-level interface, that contains generated methods.
  *
  * <p>Evolving: this interface is under development. Method signatures may change.
@@ -22,14 +27,24 @@ public interface AlertsService {
    * <p>Creates an alert. An alert is a Databricks SQL object that periodically runs a query,
    * evaluates a condition of its result, and notifies users or notification destinations if the
    * condition was met.
+   *
+   * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+   *
+   * <p>[Learn more]:
+   * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
    */
   Alert create(CreateAlert createAlert);
 
   /**
    * Delete an alert.
    *
-   * <p>Deletes an alert. Deleted alerts are no longer accessible and cannot be restored. **Note:**
+   * <p>Deletes an alert. Deleted alerts are no longer accessible and cannot be restored. **Note**:
    * Unlike queries and dashboards, alerts cannot be moved to the trash.
+   *
+   * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+   *
+   * <p>[Learn more]:
+   * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
    */
   void delete(DeleteAlertRequest deleteAlertRequest);
 
@@ -37,6 +52,11 @@ public interface AlertsService {
    * Get an alert.
    *
    * <p>Gets an alert.
+   *
+   * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+   *
+   * <p>[Learn more]:
+   * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
    */
   Alert get(GetAlertRequest getAlertRequest);
 
@@ -44,6 +64,11 @@ public interface AlertsService {
    * Get alerts.
    *
    * <p>Gets a list of alerts.
+   *
+   * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+   *
+   * <p>[Learn more]:
+   * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
    */
   Collection<Alert> list();
 
@@ -51,6 +76,11 @@ public interface AlertsService {
    * Update an alert.
    *
    * <p>Updates an alert.
+   *
+   * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+   *
+   * <p>[Learn more]:
+   * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
    */
   void update(EditAlert editAlert);
 }