+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sql/DataSourcesAPI.java
@@ -15,6 +15,11 @@ import org.slf4j.LoggerFactory;
  * <p>This API does not support searches. It returns the full list of SQL warehouses in your
  * workspace. We advise you to use any text editor, REST client, or `grep` to search the response
  * from this API for the name of your SQL warehouse as it appears in Databricks SQL.
+ *
+ * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+ *
+ * <p>[Learn more]:
+ * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
  */
 @Generated
 public class DataSourcesAPI {
@@ -38,6 +43,11 @@ public class DataSourcesAPI {
    * <p>Retrieves a full list of SQL warehouses available in this workspace. All fields that appear
    * in this API response are enumerated for clarity. However, you need only a SQL warehouse's `id`
    * to create new queries against it.
+   *
+   * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+   *
+   * <p>[Learn more]:
+   * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
    */
   public Iterable<DataSource> list() {
     return impl.list();