+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sql/DataSourcesService.java
@@ -14,6 +14,11 @@ import java.util.Collection;
  * workspace. We advise you to use any text editor, REST client, or `grep` to search the response
  * from this API for the name of your SQL warehouse as it appears in Databricks SQL.
  *
+ * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+ *
+ * <p>[Learn more]:
+ * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
+ *
  * <p>This is the high-level interface, that contains generated methods.
  *
  * <p>Evolving: this interface is under development. Method signatures may change.
@@ -26,6 +31,11 @@ public interface DataSourcesService {
    * <p>Retrieves a full list of SQL warehouses available in this workspace. All fields that appear
    * in this API response are enumerated for clarity. However, you need only a SQL warehouse's `id`
    * to create new queries against it.
+   *
+   * <p>**Note**: A new version of the Databricks SQL API will soon be available. [Learn more]
+   *
+   * <p>[Learn more]:
+   * https://docs.databricks.com/en/whats-coming.html#updates-to-the-databricks-sql-api-for-managing-queries-alerts-and-data-sources
    */
   Collection<DataSource> list();
 }